ChatBotBack/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
duckdb==1.3.0
et_xmlfile==2.0.0
Faker==19.13.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
numpy==1.24.4
packaging==25.0
pandas==1.5.3
pandasai==2.3.0
pillow==10.4.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlglot==25.34.1
sqlglotrs==0.3.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
